% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/makeTable.R
\name{collapse.results}
\alias{collapse.results}
\title{helper function to average over results from MonteCarlo() run}
\usage{
collapse.results(output, collapse)
}
\description{
helper function to average over results from MonteCarlo() run
}
\keyword{internal}

