% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeTable.R
\name{collapse_results}
\alias{collapse_results}
\title{helper function to average over results from MonteCarlo() run}
\usage{
collapse_results(output, collapse)
}
\description{
helper function to average over results from MonteCarlo() run
}
\keyword{internal}
