% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/monte carlo.R
\name{MC.inner}
\alias{MC.inner}
\title{Runs Monte Carlo. For details see documentation of wrapper function.}
\usage{
MC.inner(func, M, param.list, ret.vals, ncpus = 1, raw = FALSE,
  max.grid = 1000, packages = NULL, export.functions = NULL,
  DEBUG = FALSE)
}
\description{
Runs Monte Carlo. For details see documentation of wrapper function.
}
\keyword{internal}

