% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MergeResults.R
\name{MergeResults}
\alias{MergeResults}
\title{Merge output arrays from distributed simulation.}
\usage{
MergeResults(identifier, path, bind_along = NULL)
}
\arguments{
\item{identifier}{string that is common to the subset of files in the selected directory that are supposed to be merged.}

\item{path}{path to directory that contains the files to be merged}

\item{bind_along}{array dimension that stacks monte carlo repetitions. By default the last one.}
}
\description{
\code{MergeResults} is a utility function that allows to bind several output arrays together,
}
\details{
Arrays have to be saved using \code{save}
}
\examples{
#erg<-MergeResults(path="C:/Users/Monte Carlo Results/gg", identifier="gg", bind_along=2)
}

